#315 Refactor base protocol (#132)
Refactor the base communication protocol to support initializing and disposing a client session for a given language. The method parameters are encapsulated in dedicated parameter objects to provide a stable API. Refactor the existing `initializeServer` method to return an `InitializeResult` instead of a boolean and use a generic args map for custom arguments instead of a dedicated object. The `InitializeResult` also provides the information about which actionKinds are handled by the server. In addition, the protocol version that is implemented by the client is now passed as part of the `InitializeParameters`

Using the new initialize result  the server action handlers can be configured BEFORE the first action is dispatched in a diagram container, which means its no longer necessary to register certain early actions directly in the diagram server.

Refactor the standalone example to properly implement the new protocol.

Also:
- Remove no longer needs protocol actions for initializing/disposing a client session.
- Update webpack config of standalone example to properly support typescript source maps.


Part of eclipse-glsp/glsp/issues/315