GLSP-1063: Introduce feature modules (#267)
Introduce custom container modules that allow to specify required modules and enable conditional loading
i.e. only load if the all required modules have been loaded before.
Each feature module bind a `featureId`(symbol) when loaded which can be used by other feature modules to identify whether
a required feature module is bound.

Prime usecase for this is to facilitate `extensionModules` i.e. modules that build on top of the functionality provided by a base module (and optionally customize,rebind certain services) for certain integration scenarios (standalone, theia, vscode).

This also reduces the customization overhead e.g.  application specific integrations as it's no longer necessary to manually customize the behavior if certain default modules should not be included.
For instance if an adopter doesn't want to provide the `exportFunctionality` he can simply remove the `exportModule` and any application specific customization (`standaloneExportModule`, `theiaExportModule`) will not be loaded in the container as well.

In addition, this change is also essential if we decide to switch to a `capability` based approach similar to LSP at some point. (i.e. the server defines it`s capabilities and the client dynamically loads the modules required for this capabilities).

Fixes https://github.com/eclipse-glsp/glsp/issues/1063